+++ src/app/(home)/page.tsx
@@ -1,7 +1,7 @@
 import HomeCalendar from '@/components/_home/home-calendar';
 import HomeContent from '@/components/_home/home-content';
-import CalendarProvider from '@/lib/contexts/calendar.context';
-import DashboardProvider from '@/lib/contexts/dashboard.context';
+import CalendarProvider from '@/app/(home)/contexts/calendar.context';
+import DashboardProvider from '@/app/(home)/contexts/dashboard.context';
 
 const HomePage = async () => {
   return (
